perf(boot+pwa): Onda 4 — jssip fora do entry (−56%: 472→209KB) + precache do PWA enxuto (−21%: 6.2→4.9MB) (#17+#20)
#17 — split do contexto WebRTC: o IncomingCallModal/TransferSpikePanel
(montados no AppShellLayout = grafo estático do entry) importavam os
hooks do WebRTCCallContext.tsx, que importa SipClient → jssip (~250KB)
ESTÁTICO — o Rollup colocava o jssip inteiro no main bundle e ANULAVA o
lazy do ConditionalWebRTCProvider (módulo alcançável estaticamente não
splitta). Fix: context object + hooks + types extraídos pro módulo LEVE
src/contexts/webrtc-call-context.ts (imports só type-only); o .tsx vira
Provider-only; 5 consumidores migrados; guardrail de CI proíbe import
estático do .tsx (alias E relativo) fora de __tests__. Medido no build:
entry 472→209KB bruto; jssip 0× no entry (foi pro chunk lazy de 261KB,
staff-only pós-boot). Lente "Ver como" intacta (isLensActive segue na
fonte: makeCall/spikeTransfer/acceptIncoming); adversarial confirmou
context object 1× no dist (sem dupla instância).

#20 — precache do PWA enxuto: o SW baixava TODOS os ~260 chunks (6.2MB)
na 1ª instalação — custo real no 4G da vendedora. globIgnores SÓ em
telas inerentemente online (FarmerCopilot/voz 465KB, AdminRoutePlanner/
Leaflet js+css, posthog core 186KB, docs internas de dev); picking/
recebimento/pedido/Meu Dia/rota e todos os vendors SEGUEM precacheados
(offline-first §6.1 preservado — verificado no manifest: 8 chunks
críticos dentro). + runtimeCaching CacheFirst de /assets/ (hash no nome
= immutable): quem saiu do precache ganha offline PROGRESSIVO (visitou
1× online, fica). posthog-js ganhou nome explícito de chunk
(vendor-posthog) — o nome genérico "module-*" era inseguro de mirar.

#19 (prefetch) deliberadamente NÃO implementado: com o precache mantido
pras telas operacionais, o SW já é o prefetch — código novo seria morto.
#17b (jssip lazy) e B4 (vite:preloadError + watchdog de boot) já
estavam na main.

Revisão adversarial (subagente): split aprovado (37/37 guardrails);
3 achados corrigidos — glob UxRules com case errado (nocase só no mac;
no Linux do builder o ignore morria silencioso), chunk genérico
module-* → vendor-posthog nomeado, guardrail ampliado pra import
relativo. Rebuild + manifest re-verificado pós-fix (P1 de processo).
Validação: typecheck strict + vitest 2985/2985 + lint 0 errors.
Sem migration, sem edge.
