+++ vite.config.ts
@@ -53,6 +53,28 @@ export default defineConfig(({ mode }) => ({
       },
       workbox: {
         globPatterns: ["**/*.{js,css,html,ico,png,svg,woff2}"],
+        // Precache ENXUTO: o SW baixava TODOS os ~250 chunks (~6.2MB) na 1ª
+        // instalação e revalidava a cada deploy — custo real no 4G da
+        // vendedora. Ficam FORA do precache só chunks de telas INERENTEMENTE
+        // online (IA/voz, mapa com tiles de rede, session replay, docs
+        // internas de dev) — offline nelas não funcionaria de qualquer jeito.
+        // Telas operacionais (picking/recebimento/pedido/Meu Dia/rota) e
+        // todos os vendors continuam precacheados (offline-first, §6.1).
+        // Quem sai do precache ganha cobertura PROGRESSIVA pelo runtime
+        // caching de /assets/ (CacheFirst) abaixo: visitou 1× online, fica.
+        globIgnores: [
+          "**/FarmerCopilot-*.js", // ~465KB — copilot de voz (ElevenLabs), exige rede
+          "**/AdminRoutePlanner-*.js", // ~188KB — Leaflet; tiles do mapa exigem rede
+          "**/AdminRoutePlanner-*.css", // CSS do Leaflet (único ignorado com CSS próprio)
+          "**/vendor-posthog-*.js", // ~186KB — posthog-js core (analytics; lazy via analytics.ts)
+          "**/TechnicalDocs-*.js", // ~74KB — documentação interna de dev
+          "**/DesignSystem-*.js", // docs internas de design
+          "**/DesignPreview-*.js",
+          // ⚠️ Case EXATO do chunk (UXRules.tsx): glob é case-insensitive no
+          // macOS (nocase default do darwin) mas case-SENSITIVE no Linux do
+          // builder de produção — grafia errada = ignore morto só em prod.
+          "**/UXRules-*.js",
+        ],
         maximumFileSizeToCacheInBytes: 5 * 1024 * 1024, // 5MB
         skipWaiting: true,
         clientsClaim: true,
@@ -150,6 +172,24 @@ export default defineConfig(({ mode }) => ({
             urlPattern: /^https:\/\/.*\.supabase\.co\/(auth|realtime)/i,
             handler: "NetworkOnly",
           },
+          // ─── Chunks fora do precache: cache progressivo ─────────────
+          // Assets do Vite têm hash no nome (immutable) → CacheFirst é
+          // seguro por construção (conteúdo novo = URL nova). Cobre os
+          // chunks dos globIgnores acima (e qualquer futuro): a 1ª visita
+          // online grava no cache e a tela passa a abrir offline também.
+          // same-origin only (o pattern casa pathname em same-origin).
+          {
+            urlPattern: /\/assets\/.*\.(?:js|css)$/i,
+            handler: "CacheFirst",
+            options: {
+              cacheName: "app-assets-progressive",
+              expiration: {
+                maxEntries: 100,
+                maxAgeSeconds: 30 * 24 * 60 * 60, // 30 dias
+                purgeOnQuotaError: true,
+              },
+            },
+          },
         ],
       },
       devOptions: {
@@ -189,6 +229,11 @@ export default defineConfig(({ mode }) => ({
             if (id.includes('@tanstack/react-query')) return 'vendor-query';
             if (id.includes('@supabase/supabase-js')) return 'vendor-supabase';
             if (id.includes('recharts') || id.includes('d3-')) return 'vendor-charts';
+            // Nome explícito: sem isto o chunk do posthog-js (dynamic import
+            // em analytics.ts) nasce como "module-<hash>.js" (basename do
+            // entry ESM do pacote) — nome genérico que o globIgnores do
+            // precache não consegue mirar com segurança.
+            if (id.includes('posthog-js')) return 'vendor-posthog';
             if (id.includes('framer-motion')) return 'vendor-motion';
             if (id.includes('@radix-ui/')) return 'vendor-ui';
             if (id.includes('lucide-react')) return 'vendor-icons';
